+++ src/main/kotlin/dev/turingcomplete/intellijdevelopertoolsplugin/_internal/settings/DeveloperToolsInstanceSettings.kt
@@ -4,6 +4,7 @@ import com.intellij.credentialStore.CredentialAttributes
 import com.intellij.credentialStore.Credentials
 import com.intellij.credentialStore.generateServiceName
 import com.intellij.ide.passwordSafe.PasswordSafe
+import com.intellij.openapi.application.ApplicationManager
 import com.intellij.openapi.diagnostic.logger
 import com.intellij.ui.JBColor
 import com.intellij.util.xmlb.Converter
@@ -11,16 +12,17 @@ import com.intellij.util.xmlb.annotations.Attribute
 import com.intellij.util.xmlb.annotations.Tag
 import com.intellij.util.xmlb.annotations.XCollection
 import com.intellij.util.xmlb.annotations.XCollection.Style.v2
+import com.jetbrains.rd.util.UUID
+import com.jetbrains.rd.util.firstOrNull
 import dev.turingcomplete.intellijdevelopertoolsplugin.DeveloperToolConfiguration
-import dev.turingcomplete.intellijdevelopertoolsplugin.DeveloperToolConfiguration.PropertyContainer
+import dev.turingcomplete.intellijdevelopertoolsplugin.DeveloperToolConfiguration.PersistentProperty
 import dev.turingcomplete.intellijdevelopertoolsplugin.DeveloperToolConfiguration.PropertyType
 import dev.turingcomplete.intellijdevelopertoolsplugin.DeveloperToolConfiguration.PropertyType.CONFIGURATION
 import dev.turingcomplete.intellijdevelopertoolsplugin.DeveloperToolConfiguration.PropertyType.INPUT
 import dev.turingcomplete.intellijdevelopertoolsplugin.DeveloperToolConfiguration.PropertyType.SECRET
 import dev.turingcomplete.intellijdevelopertoolsplugin._internal.common.LocaleContainer
 import dev.turingcomplete.intellijdevelopertoolsplugin._internal.common.ValueProperty
-import java.security.Provider
-import java.security.Security
+import java.math.BigDecimal
 import java.util.*
 import java.util.concurrent.ConcurrentHashMap
 import java.util.concurrent.CopyOnWriteArrayList
@@ -31,22 +33,12 @@ internal abstract class DeveloperToolsInstanceSettings {
   // -- Properties -------------------------------------------------------------------------------------------------- //
 
   private val developerToolsConfigurations = ConcurrentHashMap<String, CopyOnWriteArrayList<DeveloperToolConfiguration>>()
+
   val lastSelectedContentNodeId: ValueProperty<String?> = ValueProperty(null)
   var expandedGroupNodeIds: MutableSet<String>? = null
     private set
 
   // -- Initialization ---------------------------------------------------------------------------------------------- //
-
-  init {
-    try {
-      val bouncyCastleProviderClass = Class.forName("org.bouncycastle.jce.provider.BouncyCastleProvider")
-      val bouncyCastleProvider = bouncyCastleProviderClass.getConstructor().newInstance()
-      Security.addProvider(bouncyCastleProvider as Provider)
-    } catch (e: Exception) {
-      log.debug("Can't load BouncyCastleProvider", e)
-    }
-  }
-
   // -- Exposed Methods --------------------------------------------------------------------------------------------- //
 
   fun setExpandedGroupNodeIds(expandedGroupNodeIds: Set<String>) {
@@ -57,7 +49,11 @@ internal abstract class DeveloperToolsInstanceSettings {
     developerToolsConfigurations[developerToolId]?.toList() ?: emptyList()
 
   fun createDeveloperToolConfiguration(developerToolId: String): DeveloperToolConfiguration {
-    val newDeveloperToolConfiguration = DeveloperToolConfiguration(name = "Workbench")
+    val newDeveloperToolConfiguration = DeveloperToolConfiguration(
+      name = "Workbench",
+      id = UUID.randomUUID(),
+      persistentProperties = emptyMap()
+    )
     developerToolsConfigurations.compute(developerToolId) { _, developerToolConfigurations ->
       (developerToolConfigurations ?: CopyOnWriteArrayList()).also { it.add(newDeveloperToolConfiguration) }
     }
@@ -71,9 +67,7 @@ internal abstract class DeveloperToolsInstanceSettings {
   open fun getState(): InstanceState {
     val stateDeveloperToolsConfigurations = developerToolsConfigurations.asSequence()
       .flatMap { (developerToolId, developerToolConfigurations) ->
-        developerToolConfigurations
-          .filter { it.properties.any { (_, property) -> property.valueChanged() } }
-          .map { createDeveloperToolConfigurationState(developerToolId, it) }
+        developerToolConfigurations.mapNotNull { createDeveloperToolConfigurationState(developerToolId, it) }
       }.toList()
 
     return InstanceState(
@@ -94,19 +88,23 @@ internal abstract class DeveloperToolsInstanceSettings {
         val developerToolConfiguration = DeveloperToolConfiguration(
           name = developerToolsConfigurationState.name!!,
           id = UUID.fromString(developerToolsConfigurationState.id!!),
-          persistentProperties = developerToolsConfigurationState.properties!!
-            .asSequence()
-            .filter { it.key != null && it.type != null }
-            .filter { it.type != INPUT || DeveloperToolsApplicationSettings.instance.saveInputs }
-            .filter { it.type != CONFIGURATION || DeveloperToolsApplicationSettings.instance.saveConfigurations }
-            .mapNotNull { property ->
+          persistentProperties = developerToolsConfigurationState.properties
+            ?.asSequence()
+            ?.filter { it.key != null && it.type != null }
+            ?.filter { it.type != INPUT || DeveloperToolsApplicationSettings.instance.saveInputs }
+            ?.filter { it.type != CONFIGURATION || DeveloperToolsApplicationSettings.instance.saveConfigurations }
+            ?.mapNotNull { property ->
               restorePropertyValue(
                 developerToolId = developerToolsConfigurationState.developerToolId!!,
                 key = property.key!!,
                 value = property.value,
                 type = property.type!!
-              )?.let { property.key!! to it }
-            }.toMap()
+              )?.let {
+                val key = property.key!!
+                key to PersistentProperty(key, it, property.type!!)
+              }
+            }
+            ?.toMap() ?: emptyMap()
         )
 
         developerToolsConfigurations.compute(developerToolsConfigurationState.developerToolId!!) { _, developerToolConfigurations ->
@@ -120,36 +118,49 @@ internal abstract class DeveloperToolsInstanceSettings {
   private fun createDeveloperToolConfigurationState(
     developerToolId: String,
     developerToolConfiguration: DeveloperToolConfiguration
-  ) = DeveloperToolConfigurationState(
-    developerToolId = developerToolId,
-    id = developerToolConfiguration.id.toString(),
-    name = developerToolConfiguration.name,
-    properties = developerToolConfiguration
-      .properties
-      .filter { (_, property) -> property.valueChanged() }
-      .filter { (_, property) -> !(!DeveloperToolsApplicationSettings.instance.saveInputs && property.type != INPUT) }
-      .filter { (_, property) -> !(!DeveloperToolsApplicationSettings.instance.saveSecrets && property.type != SECRET) }
-      .map { (key, property) ->
-        storeProperty(developerToolId = developerToolId, key = key, property = property)
-      }
-  )
+  ): DeveloperToolConfigurationState? {
+    val properties = if (developerToolConfiguration.wasConsumedByDeveloperTool) {
+      developerToolConfiguration.properties
+        .filter { (_, property) -> property.valueIsNotDefaultOrExample() }
+        .map { PersistentProperty(key = it.key, value = it.value.reference.get(), it.value.type) }
+    }
+    else {
+      developerToolConfiguration.persistentProperties.values
+    }
+    if (properties.isEmpty()) {
+      return null
+    }
+
+    return DeveloperToolConfigurationState(
+      developerToolId = developerToolId,
+      id = developerToolConfiguration.id.toString(),
+      name = developerToolConfiguration.name,
+      properties = properties
+        .filter { !(!DeveloperToolsApplicationSettings.instance.saveInputs && it.type != INPUT) }
+        .filter { !(!DeveloperToolsApplicationSettings.instance.saveSecrets && it.type != SECRET) }
+        .map { storeProperty(developerToolId = developerToolId, key = it.key, property = it) }
+    )
+  }
 
   private fun storeProperty(
     developerToolId: String,
     key: String,
-    property: PropertyContainer
+    property: PersistentProperty
   ): DeveloperToolConfigurationProperty {
     return when (property.type) {
       CONFIGURATION, INPUT ->
-        DeveloperToolConfigurationProperty(key = key, value = property.reference.get(), type = property.type)
+        DeveloperToolConfigurationProperty(key = key, value = property.value, type = property.type)
 
       SECRET -> {
         val credentialAttribute = createPropertyCredentialAttribute(developerToolId = developerToolId, propertyKey = key)
-        PasswordSafe.instance.set(credentialAttribute, Credentials(null, property.reference.get() as String))
+        ApplicationManager.getApplication().executeOnPooledThread {
+          // Not allowed to be executed on the EDT
+          PasswordSafe.instance.set(credentialAttribute, Credentials(null, property.value as String))
+        }
         // The value is not stored in the XML file, but we still need the
         // `DeveloperToolConfigurationProperty` to represent the property.
-        // Therefore we use `*******` as a dummy value.
-        DeveloperToolConfigurationProperty(key = key, value = "*******", type = property.type)
+        // Therefore, we use `SAVED_IN_KEYSTORE` as a placeholder value.
+        DeveloperToolConfigurationProperty(key = key, value = "SAVED_IN_KEYSTORE", type = property.type)
       }
     }
   }
@@ -250,20 +261,21 @@ internal abstract class DeveloperToolsInstanceSettings {
         is String -> value to String::class.qualifiedName!!
         is JBColor -> value.rgb.toString() to JBColor::class.qualifiedName!!
         is LocaleContainer -> value.locale.toLanguageTag() to LocaleContainer::class.qualifiedName!!
+        is BigDecimal -> value.toString() to BigDecimal::class.qualifiedName!!
         else -> error("Unsupported configuration property type: ${value::class.qualifiedName}")
       }
       return "${valueType}$PROPERTY_TYPE_VALUE_DELIMITER$serializedValue"
     }
 
     /**
-     * If the value can't be restored this method must return null. All
-     * properties with null values will be filtered in
+     * If the value can't be restored, this method must return null.
+     * All properties with null values will be filtered in
      * [DeveloperToolsInstanceSettings.loadState].
      */
     override fun fromString(serializedValue: String): Any? {
       val valueAndType = serializedValue.split(PROPERTY_TYPE_VALUE_DELIMITER, limit = 2)
       check(valueAndType.size == 2) { "Malformed serialized value: $serializedValue" }
-      val valueType = valueAndType[0]
+      val valueType = applyTypeLegacy(valueAndType[0])
       val value = valueAndType[1]
       return when (valueType) {
         Boolean::class.qualifiedName -> value.toBoolean()
@@ -274,8 +286,16 @@ internal abstract class DeveloperToolsInstanceSettings {
         String::class.qualifiedName -> value
         JBColor::class.qualifiedName -> JBColor(value.toInt(), value.toInt())
         LocaleContainer::class.qualifiedName -> LocaleContainer(Locale.forLanguageTag(value))
+        BigDecimal::class.qualifiedName -> BigDecimal(value)
         else -> parseValue(valueType, value)
       }
+    }
+
+    private fun applyTypeLegacy(type: String): String {
+      val legacyToApply = pre320TypePackageLegacy.filter { type.startsWith(it.key) }.firstOrNull()
+        ?: return type
+
+      return "${legacyToApply.value}${type.substring(legacyToApply.key.length)}"
     }
 
     private fun parseValue(valueType: String, value: String): Any? {
@@ -288,7 +308,7 @@ internal abstract class DeveloperToolsInstanceSettings {
           error("Unsupported configuration property: $valueType")
         }
       } catch (e: Exception) {
-        log.error("Failed to load class '$valueType' of value '$value'", e)
+        log.warn("Failed to load class '$valueType' of value '$value'", e)
         null
       }
     }
@@ -300,6 +320,10 @@ internal abstract class DeveloperToolsInstanceSettings {
 
     private val log = logger<DeveloperToolsInstanceSettings>()
 
+    private val pre320TypePackageLegacy = mapOf(
+      "dev.turingcomplete.intellijdevelopertoolsplugins._internal.tool." to "dev.turingcomplete.intellijdevelopertoolsplugin._internal.tool.ui.",
+      "dev.turingcomplete.intellijdevelopertoolsplugins._internal.common." to "dev.turingcomplete.intellijdevelopertoolsplugin._internal.common."
+    )
     private const val PROPERTY_TYPE_VALUE_DELIMITER = "|"
     private val SUPPORTED_TYPES = setOf<KClass<*>>(
       Boolean::class,
@@ -310,6 +334,7 @@ internal abstract class DeveloperToolsInstanceSettings {
       String::class,
       JBColor::class,
       LocaleContainer::class,
+      BigDecimal::class
     )
 
     fun <T : Any> assertPersistableType(type: KClass<T>, propertyType: PropertyType): KClass<T> {