+++ CHANGELOG.md
@@ -10,6 +10,19 @@
 
 ### Fixed
 
+## 3.3.0 - 2024-01-15
+
+### Added
+
+- Add time conversion tool
+- Add 'What's New' overview to the main menu
+
+### Fixed
+
+- Inputs and configurations for tools that were not opened were lost the next time the settings were saved.
+- Loading the tool window settings causes an error if one of the tools has a saved secret.
+- In the JWT Encoder/Decoder the "Public Key" field was visible for the HMAC signature algorithm.
+
 ## 3.2.0 - 2024-01-06
 
 ### Added